[IMP] point_of_sale: print saleDetailsReport from closePosPopup
In this commit we make it possible for a user to
print the Sales Report of the current session by
clicking at the "print" icon at the lower right
corner of the closing pos popup. This is a commonly
used feature. Thus, we make the UI more usable.

Note that this button will be displayed only when
a printer is connected.

This functionality already existed at the saleDetailsButton.
So in order to make the code reusable we moved the logic of
the saleDetailsButton to the Chrome parent component. This
way we expose the feature from the ClosePosPopup as well.

After moving the logic, the saleDetailsButton class remained
empty, so we removed it.

task-2669747
