--- addons/point_of_sale/static/src/xml/ChromeWidgets/SaleDetailsButton.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates id="template" xml:space="preserve">
-
-    <t t-name="SaleDetailsButton" owl="1">
-        <div class="oe_status">
-            <div class="js_connected oe_icon">
-                <i class="fa fa-fw fa-print" role="img" aria-label="Print" t-on-click="onClick"
-                   title="Print a report with all the sales of the current PoS Session"></i>
-            </div>
-        </div>
-    </t>
-
-</templates>
