+++ addons/point_of_sale/static/src/xml/Popups/ClosePosPopup.xml
@@ -83,10 +83,15 @@
                     <div class="button highlight" t-on-click="onCloseSession">Close Session</div>
                     <div class="button" t-on-click="closePos" title="Visit the Backend but keep session open">Backend</div>
                     <div class="button" t-att-class="{'disabled': !canCancel()}" t-on-click="cancelPopup">Discard</div>
-                    <div class="small button icon"><i class="fa fa-print" role="img" title="Print the report"/></div>
-                    <div class="small button icon" t-on-click="downloadSalesReport" 
+                    <div class="small button icon" 
+                        t-on-click="downloadSalesReport" 
                         title="Download a report with all the sales of the current PoS Session">
                             <i class="fa fa-download" role="img"/>
+                    </div>
+                    <div class="small button icon" t-if="env.proxy.printer" 
+                        t-on-click="() => this.trigger('print-sale-details')"
+                        title="Print a report with all the sales of the current PoS Session">
+                            <i class="fa fa-print" role="img"/>
                     </div>
                 </footer>
             </div>
